Release v2.0.0

+++ doc/basics/multisampling/shaders/multisampler.frag
@@ -4,7 +4,7 @@ precision highp float;
 uniform highp sampler2DMS textureSampler;
 uniform highp int sampleCount;
 
-in lowp vec2 v_TextureCoordinate;
+in highp vec2 v_TextureCoordinate;
 out vec4 fragColor;
 
 void main(void)
